% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R, R/cubes.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{obj_sum}
\alias{as_tibble}
\alias{slice}
\alias{select}
\alias{relocate}
\alias{tbl_format_setup}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{relocate}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{slice}}}

  \item{pillar}{\code{\link[pillar:type_sum]{obj_sum}}, \code{\link[pillar]{tbl_format_setup}}}

  \item{tibble}{\code{\link[tibble]{as_tibble}}}
}}

